Check module enabled for currencies migration

+++ src/Database/Migrations/2021_10_23_082414_create_currencies_table.php
@@ -13,18 +13,20 @@ class CreateCurrenciesTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create(config('world.migrations.currencies.table_name'), function (Blueprint $table) {
-			$table->id();
-			$table->foreignId('country_id');
-			$table->string('name');
-			$table->string('code');
-			$table->tinyInteger('precision')->default(2);
-			$table->string('symbol');
-			$table->string('symbol_native');
-			$table->tinyInteger('symbol_first')->default(1);
-			$table->string('decimal_mark', 1)->default('.');
-			$table->string('thousands_separator', 1)->default(',');
-		});
+		if(config('world.modules.currencies')){
+			Schema::create(config('world.migrations.currencies.table_name'), function (Blueprint $table) {
+				$table->id();
+				$table->foreignId('country_id');
+				$table->string('name');
+				$table->string('code');
+				$table->tinyInteger('precision')->default(2);
+				$table->string('symbol');
+				$table->string('symbol_native');
+				$table->tinyInteger('symbol_first')->default(1);
+				$table->string('decimal_mark', 1)->default('.');
+				$table->string('thousands_separator', 1)->default(',');
+			});
+		}
 	}
 
 	/**
@@ -34,6 +36,8 @@ class CreateCurrenciesTable extends Migration
 	 */
 	public function down()
 	{
-		Schema::dropIfExists(config('world.migrations.currencies.table_name'));
+		if(config('world.modules.currencies')){
+			Schema::dropIfExists(config('world.migrations.currencies.table_name'));
+		}
 	}
 }
